+++ .gitignore
@@ -20,7 +20,7 @@ erl_crash.dump
 *.ez
 
 # Ignore package tarball (built via "mix hex.build").
-petal_boilerplate-*.tar
+eblox-*.tar
 
 # Ignore assets that are produced by build tools.
 /priv/static/assets/
