+++ config/prod.exs
@@ -9,7 +9,7 @@ import Config
 # manifest is generated by the `mix phx.digest` task,
 # which you should run after static files are built and
 # before starting your production server.
-config :petal_boilerplate, PetalBoilerplateWeb.Endpoint,
+config :eblox, EbloxWeb.Endpoint,
   cache_static_manifest: "priv/static/cache_manifest.json"
 
 # Do not print debug messages in production
@@ -20,7 +20,7 @@ config :logger, level: :info
 # To get SSL working, you will need to add the `https` key
 # to the previous section and set your `:url` port to 443:
 #
-#     config :petal_boilerplate, PetalBoilerplateWeb.Endpoint,
+#     config :eblox, EbloxWeb.Endpoint,
 #       ...,
 #       url: [host: "example.com", port: 443],
 #       https: [
@@ -44,7 +44,7 @@ config :logger, level: :info
 # We also recommend setting `force_ssl` in your endpoint, ensuring
 # no data is ever sent via http, always redirecting to https:
 #
-#     config :petal_boilerplate, PetalBoilerplateWeb.Endpoint,
+#     config :eblox, EbloxWeb.Endpoint,
 #       force_ssl: [hsts: true]
 #
 # Check `Plug.SSL` for all available options in `force_ssl`.
