+++ config/runtime.exs
@@ -9,7 +9,7 @@ import Config
 
 # Start the phoenix server if environment is set and running in a release
 if System.get_env("PHX_SERVER") && System.get_env("RELEASE_NAME") do
-  config :petal_boilerplate, PetalBoilerplate.Endpoint, server: true
+  config :eblox, Eblox.Endpoint, server: true
 end
 
 if config_env() == :prod do
@@ -22,7 +22,7 @@ if config_env() == :prod do
 
   maybe_ipv6 = if System.get_env("ECTO_IPV6"), do: [:inet6], else: []
 
-  config :petal_boilerplate, PetalBoilerplate.Repo,
+  config :eblox, Eblox.Repo,
     # ssl: true,
     url: database_url,
     pool_size: String.to_integer(System.get_env("POOL_SIZE") || "10"),
@@ -43,7 +43,7 @@ if config_env() == :prod do
   host = System.get_env("PHX_HOST") || "example.com"
   port = String.to_integer(System.get_env("PORT") || "4000")
 
-  config :petal_boilerplate, PetalBoilerplateWeb.Endpoint,
+  config :eblox, EbloxWeb.Endpoint,
     url: [host: host, port: 443],
     http: [
       # Enable IPv6 and bind on all interfaces.
@@ -60,7 +60,7 @@ if config_env() == :prod do
   # If you are doing OTP releases, you need to instruct Phoenix
   # to start each relevant endpoint:
   #
-  #     config :petal_boilerplate, PetalBoilerplateWeb.Endpoint, server: true
+  #     config :eblox, EbloxWeb.Endpoint, server: true
   #
   # Then you can assemble a release by calling `mix release`.
   # See `mix help release` for more information.
@@ -71,7 +71,7 @@ if config_env() == :prod do
   # Also, you may need to configure the Swoosh API client of your choice if you
   # are not using SMTP. Here is an example of the configuration:
   #
-  #     config :petal_boilerplate, PetalBoilerplate.Mailer,
+  #     config :eblox, Eblox.Mailer,
   #       adapter: Swoosh.Adapters.Mailgun,
   #       api_key: System.get_env("MAILGUN_API_KEY"),
   #       domain: System.get_env("MAILGUN_DOMAIN")
