+++ lib/eblox/mailer.ex
@@ -1,0 +1,3 @@
+defmodule Eblox.Mailer do
+  use Swoosh.Mailer, otp_app: :eblox
+end
