+++ lib/eblox/repo.ex
@@ -1,0 +1,5 @@
+defmodule Eblox.Repo do
+  use Ecto.Repo,
+    otp_app: :eblox,
+    adapter: Ecto.Adapters.Postgres
+end
