+++ lib/eblox_web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule EbloxWeb.PageController do
+  use EbloxWeb, :controller
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+end
