+++ lib/eblox_web/views/layout_view.ex
@@ -1,0 +1,7 @@
+defmodule EbloxWeb.LayoutView do
+  use EbloxWeb, :view
+
+  # Phoenix LiveDashboard is available only in development by default,
+  # so we instruct Elixir to not warn if the dashboard route is missing.
+  @compile {:no_warn_undefined, {Routes, :live_dashboard_path, 2}}
+end
