+++ lib/eblox_web/views/page_view.ex
@@ -1,0 +1,3 @@
+defmodule EbloxWeb.PageView do
+  use EbloxWeb, :view
+end
