--- lib/petal_boilerplate/mailer.ex
@@ -1,3 +1,0 @@
-defmodule PetalBoilerplate.Mailer do
-  use Swoosh.Mailer, otp_app: :petal_boilerplate
-end
