--- lib/petal_boilerplate/repo.ex
@@ -1,5 +1,0 @@
-defmodule PetalBoilerplate.Repo do
-  use Ecto.Repo,
-    otp_app: :petal_boilerplate,
-    adapter: Ecto.Adapters.Postgres
-end
