--- lib/petal_boilerplate_web/controllers/page_controller.ex
@@ -1,7 +1,0 @@
-defmodule PetalBoilerplateWeb.PageController do
-  use PetalBoilerplateWeb, :controller
-
-  def index(conn, _params) do
-    render(conn, "index.html")
-  end
-end
