--- lib/petal_boilerplate_web/templates/layout/root.html.heex
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class={if assigns[:color_scheme] == "dark", do: "dark", else: ""}>
-  <head>
-    <meta charset="utf-8"/>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <%= csrf_meta_tag() %>
-    <%= live_title_tag assigns[:page_title] || "PetalBoilerplate", suffix: " · Phoenix Framework" %>
-    <link phx-track-static rel="stylesheet" href={Routes.static_path(@conn, "/assets/app.css")}/>
-    <script defer src="https://unpkg.com/@alpinejs/collapse@3.x.x/dist/cdn.min.js"></script>
-    <script defer src="https://unpkg.com/alpinejs@3.x.x/dist/cdn.min.js"></script>
-
-    <!-- Used in css-theme-switch.js - remove if you don't support light/dark mode -->
-    <script defer src="https://unpkg.com/js-cookie@3.x.x/dist/js.cookie.min.js"></script>
-
-    <script defer phx-track-static type="text/javascript" src={Routes.static_path(@conn, "/assets/app.js")}></script>
-  </head>
-  <body>
-    <%= @inner_content %>
-  </body>
-</html>
