--- lib/petal_boilerplate_web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule PetalBoilerplateWeb.PageView do
-  use PetalBoilerplateWeb, :view
-end
