+++ test/eblox_web/controllers/page_controller_test.exs
@@ -1,0 +1,8 @@
+defmodule EbloxWeb.PageControllerTest do
+  use EbloxWeb.ConnCase
+
+  test "GET /", %{conn: conn} do
+    conn = get(conn, "/")
+    assert html_response(conn, 200) =~ "Petal"
+  end
+end
