+++ test/eblox_web/views/page_view_test.exs
@@ -1,0 +1,3 @@
+defmodule EbloxWeb.PageViewTest do
+  use EbloxWeb.ConnCase, async: true
+end
