--- test/petal_boilerplate_web/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule PetalBoilerplateWeb.PageControllerTest do
-  use PetalBoilerplateWeb.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get(conn, "/")
-    assert html_response(conn, 200) =~ "Petal"
-  end
-end
