--- test/petal_boilerplate_web/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule PetalBoilerplateWeb.PageViewTest do
-  use PetalBoilerplateWeb.ConnCase, async: true
-end
