+++ test/support/conn_case.ex
@@ -1,4 +1,4 @@
-defmodule PetalBoilerplateWeb.ConnCase do
+defmodule EbloxWeb.ConnCase do
   @moduledoc """
   This module defines the test case to be used by
   tests that require setting up a connection.
@@ -11,7 +11,7 @@ defmodule PetalBoilerplateWeb.ConnCase do
   we enable the SQL sandbox, so changes done to the database
   are reverted at the end of every test. If you are using
   PostgreSQL, you can even run database tests asynchronously
-  by setting `use PetalBoilerplateWeb.ConnCase, async: true`, although
+  by setting `use EbloxWeb.ConnCase, async: true`, although
   this option is not recommended for other databases.
   """
 
@@ -22,17 +22,17 @@ defmodule PetalBoilerplateWeb.ConnCase do
       # Import conveniences for testing with connections
       import Plug.Conn
       import Phoenix.ConnTest
-      import PetalBoilerplateWeb.ConnCase
+      import EbloxWeb.ConnCase
 
-      alias PetalBoilerplateWeb.Router.Helpers, as: Routes
+      alias EbloxWeb.Router.Helpers, as: Routes
 
       # The default endpoint for testing
-      @endpoint PetalBoilerplateWeb.Endpoint
+      @endpoint EbloxWeb.Endpoint
     end
   end
 
   setup tags do
-    pid = Ecto.Adapters.SQL.Sandbox.start_owner!(PetalBoilerplate.Repo, shared: not tags[:async])
+    pid = Ecto.Adapters.SQL.Sandbox.start_owner!(Eblox.Repo, shared: not tags[:async])
     on_exit(fn -> Ecto.Adapters.SQL.Sandbox.stop_owner(pid) end)
     {:ok, conn: Phoenix.ConnTest.build_conn()}
   end
